Repeating the same experiments

+++ src/knobab/server/tables/ActTable.cpp
@@ -165,7 +165,7 @@ void ActTable::indexing2() { // todo: rename as indexing, and remove expectedOrd
 //    }
 //}
 
-std::pair<uint32_t, uint32_t> ActTable::resolve_index(act_t id) const {
+std::pair<size_t, size_t> ActTable::resolve_index(act_t id) const {
     if (primary_index.size() < id)
         return {-1, -1};
     else {
